hints & paths

+++ Indicators/Indics_Normalization.py
@@ -1,7 +1,7 @@
 import numexpr as ne # type: ignore
 import numpy as np
 from Metrics import rolling_mean, rolling_median, rolling_min, rolling_max, rolling_volatility
-from Utilitary import ArrayFloat, Float32
+from Utilitary import ArrayFloat
 import numbagg as nb
 
 def bfill(array: ArrayFloat) -> ArrayFloat:
@@ -30,7 +30,7 @@ def rolling_median_normalisation(
     limit:int = 1
     ) -> ArrayFloat:
 
-    adjusted_signal_array = np.empty_like(signal_array, dtype=Float32)
+    adjusted_signal_array = np.empty_like(signal_array)
     dict = {
         "signal_array": signal_array,
         "median_array": rolling_median(signal_array, length=window_length, min_length=window_length),
@@ -51,7 +51,7 @@ def rolling_std_normalisation(
     limit:int = 1
     ) -> ArrayFloat:
 
-    adjusted_signal_array = np.empty_like(signal_array, dtype=Float32)
+    adjusted_signal_array = np.empty_like(signal_array)
     dict = {
         "signal_array": signal_array,
         "median_array": rolling_median(signal_array, length=window_length, min_length=window_length),
